+++ gdsfactory/tests/test_get_bundle_from_waypoints.py
@@ -10,7 +10,6 @@ from gdsfactory.routing.get_bundle_from_waypoints import get_bundle_from_waypoin
 def test_get_bundle_from_waypointsB(
     data_regression: DataRegressionFixture,
     check: bool = True,
-    layer: gf.types.Layer = (1, 0),
 ) -> Component:
 
     ys1 = np.array([0, 5, 10, 15, 30, 40, 50, 60]) + 0.0
@@ -18,7 +17,7 @@ def test_get_bundle_from_waypointsB(
     N = ys1.size
 
     ports1 = [
-        Port(name=f"A_{i}", midpoint=(0, ys1[i]), width=0.5, orientation=0, layer=layer)
+        Port(name=f"A_{i}", midpoint=(0, ys1[i]), width=0.5, orientation=0)
         for i in range(N)
     ]
     ports2 = [
@@ -27,7 +26,6 @@ def test_get_bundle_from_waypointsB(
             midpoint=(500, ys2[i]),
             width=0.5,
             orientation=180,
-            layer=layer,
         )
         for i in range(N)
     ]
@@ -58,7 +56,6 @@ def test_get_bundle_from_waypointsB(
 def test_get_bundle_from_waypointsC(
     data_regression: DataRegressionFixture,
     check: bool = True,
-    layer: gf.types.Layer = (1, 0),
 ) -> Component:
 
     ys1 = np.array([0, 5, 10, 15, 20, 60, 70, 80, 120, 125])
@@ -66,7 +63,7 @@ def test_get_bundle_from_waypointsC(
     N = ys1.size
 
     ports1 = [
-        Port(name=f"A_{i}", midpoint=(0, ys1[i]), width=0.5, orientation=0, layer=layer)
+        Port(name=f"A_{i}", midpoint=(0, ys1[i]), width=0.5, orientation=0)
         for i in range(N)
     ]
     ports2 = [
@@ -75,7 +72,6 @@ def test_get_bundle_from_waypointsC(
             midpoint=(600, ys2[i]),
             width=0.5,
             orientation=180,
-            layer=layer,
         )
         for i in range(N)
     ]
@@ -103,7 +99,42 @@ def test_get_bundle_from_waypointsC(
     return c
 
 
+def test_get_bundle_from_waypoints_staggered(
+    data_regression: DataRegressionFixture,
+    check: bool = True,
+):
+
+    c = gf.Component()
+    r = c << gf.c.array(component=gf.c.straight)
+    r.movex(60)
+    r.movey(40)
+
+    lt = c << gf.c.straight(length=15)
+    lb = c << gf.c.straight(length=5)
+    lt.movey(5)
+
+    ports1 = lt.get_ports_list(orientation=0) + lb.get_ports_list(orientation=0)
+    ports2 = r.get_ports_list(orientation=180)
+
+    dx = 20
+    p0 = ports1[0].midpoint + (dx, 0)
+    p1 = (ports1[0].midpoint[0] + dx, ports2[0].midpoint[1])
+    waypoints = (p0, p1)
+
+    routes = gf.routing.get_bundle_from_waypoints(ports1, ports2, waypoints=waypoints)
+    lengths = {}
+    for i, route in enumerate(routes):
+        c.add(route.references)
+        lengths[i] = route.length
+
+    if check:
+        data_regression.check(lengths)
+
+    return c
+
+
 if __name__ == "__main__":
     # c = test_get_bundle_from_waypointsC(None, check=False)
-    c = test_get_bundle_from_waypointsB(None, check=False)
+    # c = test_get_bundle_from_waypointsB(None, check=False)
+    c = test_get_bundle_from_waypoints_staggered(None, check=False)
     c.show()